+++ src/main/java/http/commands/HttpRequestCommand.java
@@ -1,6 +1,6 @@
 package http.commands;
 
-import http.HttpResponse;
+import http.response.HttpResponse;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
