+++ src/main/java/http/request/RequestStartLine.java
@@ -1,10 +1,10 @@
-package http;
+package http.request;
 
 import http.HttpMethod;
 
 public class RequestStartLine {
-    private HttpMethod httpMethod;
-    private String url;
+    private final HttpMethod httpMethod;
+    private final String url;
 
     public RequestStartLine(HttpMethod httpMethod, String url) {
         this.httpMethod = httpMethod;
