+++ src/main/java/http/response/ResponseStartLine.java
@@ -1,4 +1,4 @@
-package http;
+package http.response;
 
 import http.HttpStatus;
 import http.HttpVersion;
